完善学生能力详细分析逻辑

+++ src/main/java/com/shangshufang/apiservice/entity/StudentExercisePercentAnalysisEntity.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class StudentExercisePercentAnalysisEntity {
     private String months;
-    private int finishPercent;
-    private int onceCompilationSuccessPercent;
-    private int onceRunCorrectPercent;
+    private float finishPercent;
+    private float onceCompilationSuccessPercent;
+    private float onceRunCorrectPercent;
 }
