+++ src/main/java/com/shangshufang/apiservice/vo/StudentExercisePercentAnalysisVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class StudentExercisePercentAnalysisVO {
     private String months;
-    private int finishPercent;
-    private int onceCompilationSuccessPercent;
-    private int onceRunCorrectPercent;
+    private float finishPercent;
+    private float onceCompilationSuccessPercent;
+    private float onceRunCorrectPercent;
 }
